+++ src/client/CookingAppFE/components/screens/UserMenu.js
@@ -1,21 +1,38 @@
-import React from "react";
-import { View, Text, TouchableOpacity } from "react-native";
+import React, { useState } from "react";
+import { View, Text, TouchableOpacity, Image, Alert } from "react-native";
 import Ionicons from "react-native-vector-icons/Ionicons";
 import tw from "twrnc";
 import { useNavigation } from "@react-navigation/native";
 import { useSelector } from "react-redux";
+import * as ImagePicker from "expo-image-picker";
 
 const UserMenu = () => {
   const navigation = useNavigation();
   const isDarkTheme = useSelector(state => state.ui.isDarkTheme);
+  const [profileImage, setProfileImage] = useState(null);
+
+  const pickImage = async () => {
+    let result = await ImagePicker.launchImageLibraryAsync({
+      mediaTypes: ImagePicker.MediaTypeOptions.Images,
+      allowsEditing: true,
+      aspect: [4, 3],
+      quality: 1
+    });
+
+    if (!result.canceled) {
+      setProfileImage(result.uri);
+    } else {
+      Alert.alert("Error", "Image selection was cancelled");
+    }
+  };
 
   return (
     <View style={tw`flex-1 ${isDarkTheme ? "bg-[#202020]" : "bg-white"} p-4`}>
       <View style={tw`flex-row justify-center items-center mb-4`}>
-        <TouchableOpacity onPress={() => navigation.navigate("ProfileSettings")}>
+        <TouchableOpacity onPress={pickImage}>
           <Ionicons name="person-circle" size={50} color={isDarkTheme ? "white" : "black"} />
         </TouchableOpacity>
-
+        {profileImage && <Image source={{ uri: profileImage }} style={tw`w-32 h-32 rounded-full`} />}
         <TouchableOpacity onPress={() => console.log("Close menu")}>
           <Text style={tw`text-2xl font-bold ${isDarkTheme ? "text-white" : "text-black"}`}></Text>
         </TouchableOpacity>
